+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Added single select field form option validation.
 * Changed all cookies to SameSite=lax.
 * Fixed the "Ignored attempt to cancel a touchmove" error.